Add theming support behind a feature gate
---

At a high level, this commit is the starting point for dual light/dark
theme support. By design it aims to leave dark mode untouched, while
adding a rudimentary (palette-swaps-only) implementation of light mode.

This entails a few core additions:
* A new dependency on "next-themes"
* A new ThemeProvider wrapper to apply the theme selection globally
* A new global action button to toggle between light and dark modes
* A new localstorage check that gates access to said button

To enable dynamic themes, a few core css & tailwind edits are in place:
* Existing usages of the default colors have been made explicit in css
* Light mode equivalents have been added (reversed sub-number values)
* Both of the above are exposed as shared css custom properties
* Tailwind references these css properties for its color palettes

The remainder of this commit's edits fall into one of a few categories:
* Explicit black/white usages have been aliased to solid background/surface
* Existing light-only (non-"dark:") selectors have been removed
* A few implied dark elements made explicit (black archetype backgrounds)