Convert toggle button to theme select menu
In brief, the dark/light toggle is now replaced with
a dialog list of themes to select from.
Theme changes are instant, allowing for "preview"
without leaving the modal.

A variety of changes enable this:
* The tailwind darkMode selector is aware of data-theme
* The core css vars are renamed to match the palette
* This includes renaming hardcoded c/v/y usages
* The per-theme css blocks utilize data-theme selectors
* The screenshot forces a (default) dark data-theme

This has the immediate benefit of allowing users to
reset their preference to follow their system
after trying out the toggle button beforehand.

This commit also includes a handful of new
custom themes as a proof-of-concept.
These themes are copy/pastes (with interpolation)
of established color schemes.
In order to allow for color and contrast refinement,
themes can be set to live case-by-case.
Users can opt in to trying out all non-live (beta)
themes via "r2tk-themes-beta" in localstorage.