Add a placeholder for the theme button
Since the button is dynamicaly loaded, there is a
slight delay before it renders. This pop-in was a
minor annoyance before, but with the expandable
GAB menu it is now more apparent. Attempting to
expand these immediately after opening a page
will result in the two non-dynamic icons sliding
into the leftmost slots, then immediately snapping
to the right when the theme button is loaded.

This commit adds a placeholder theme button that
is visually identical, but non-functional
(and therefore ssr-friendly) for use during this
brief period.