Eliminate layout shift on dialog/menu change
With the way the footer works, every page has a scrollbar.
Additionally, UI components such as dialogs and dropdown menus have accessibility
features such as scroll locking enabled by default.
This means that the scrollbar is hidden on interaction, which leads to many
elements on the pages undergoing a jarring layout shift.

This commit adds a right-side-only gutter for the scrollbar.
It then makes a few css tweaks to the navbar and GAB to align them appropriately.