Reworked Mudtooth's Snake Oil logic
Previously, Mudtooth's Snake Oil could be equipped as long as it was not the last concoction slot. It would add another concoction slot to offset the slot it took. This is not exactly how it should work. Rather, it shoudl always be equippable. Added new logic and a button allowing you to add Mudtooth's Snake Oil at any time, even if all concoction slots are full.