BUGFIX: Rusty weapon handling
I had labelled Shovel as a rusty item earlier to fix an issue with quality build check. The bigger issue is that I forgot melee weapons can only have mods equipped if the mods are linked to the melee weapon; you cannot add a melee mod. This also introduced a bug in the builder when I did the previous fix, which would allow you to add mods to melee weapons that shouldn't allow it. I reverted that change, and updated the quality build check to determine the correct required mod total based on the melee weapon equipped.